+++ src/runtime/importhook.cs
@@ -23,27 +23,84 @@ namespace Python.Runtime
                 module_def = ModuleDefOffset.AllocModuleDef("clr");
             }
         }
+
+        internal static void ReleaseModuleDef()
+        {
+            if (module_def == IntPtr.Zero)
+            {
+                return;
+            }
+            ModuleDefOffset.FreeModuleDef(module_def);
+            module_def = IntPtr.Zero;
+        }
 #endif
+
+        /// <summary>
+        /// Get a <i>New reference</i> to the builtins module.
+        /// </summary>
+        static IntPtr GetNewRefToBuiltins()
+        {
+            if (Runtime.IsPython3)
+            {
+                return Runtime.PyImport_ImportModule("builtins");
+            }
+            else
+            {
+                // dict is a borrowed ref, no need to decref
+                IntPtr dict = Runtime.PyImport_GetModuleDict();
+
+                // GetItemString is a borrowed ref; incref to get a new ref
+                IntPtr builtins = Runtime.PyDict_GetItemString(dict, "__builtin__");
+                Runtime.XIncref(builtins);
+                return builtins;
+            }
+        }
+
+        /// <summary>
+        /// Initialize just the __import__ hook itself.
+        /// </summary>
+        static void InitImport()
+        {
+            // We replace the built-in Python __import__ with our own: first
+            // look in CLR modules, then if we don't find any call the default
+            // Python __import__.
+            IntPtr builtins = GetNewRefToBuiltins();
+            py_import = Runtime.PyObject_GetAttrString(builtins, "__import__");
+            PythonException.ThrowIfIsNull(py_import);
+
+            hook = new MethodWrapper(typeof(ImportHook), "__import__", "TernaryFunc");
+            int res = Runtime.PyObject_SetAttrString(builtins, "__import__", hook.ptr);
+            PythonException.ThrowIfIsNotZero(res);
+
+            Runtime.XDecref(builtins);
+        }
+
+        /// <summary>
+        /// Restore the __import__ hook.
+        /// </summary>
+        static void RestoreImport()
+        {
+            IntPtr builtins = GetNewRefToBuiltins();
+
+            int res = Runtime.PyObject_SetAttrString(builtins, "__import__", py_import);
+            PythonException.ThrowIfIsNotZero(res);
+            Runtime.XDecref(py_import);
+            py_import = IntPtr.Zero;
+
+            hook.Release();
+            hook = null;
+
+            Runtime.XDecref(builtins);
+        }
 
         /// <summary>
         /// Initialization performed on startup of the Python runtime.
         /// </summary>
         internal static void Initialize()
         {
-            // Initialize the Python <--> CLR module hook. We replace the
-            // built-in Python __import__ with our own. This isn't ideal,
-            // but it provides the most "Pythonic" way of dealing with CLR
-            // modules (Python doesn't provide a way to emulate packages).
-            IntPtr dict = Runtime.PyImport_GetModuleDict();
+            InitImport();
 
-            IntPtr mod = Runtime.IsPython3
-                ? Runtime.PyImport_ImportModule("builtins")
-                : Runtime.PyDict_GetItemString(dict, "__builtin__");
-
-            py_import = Runtime.PyObject_GetAttrString(mod, "__import__");
-            hook = new MethodWrapper(typeof(ImportHook), "__import__", "TernaryFunc");
-            Runtime.PyObject_SetAttrString(mod, "__import__", hook.ptr);
-
+            // Initialize the clr module and tell Python about it.
             root = new CLRModule();
 
 #if PYTHON3
@@ -61,6 +118,7 @@ namespace Python.Runtime
             Runtime.XIncref(root.pyHandle); // we are using the module two times
             py_clr_module = root.pyHandle; // Alias handle for PY2/PY3
 #endif
+            IntPtr dict = Runtime.PyImport_GetModuleDict();
             Runtime.PyDict_SetItemString(dict, "CLR", py_clr_module);
             Runtime.PyDict_SetItemString(dict, "clr", py_clr_module);
         }
@@ -75,24 +133,21 @@ namespace Python.Runtime
             {
                 return;
             }
-            IntPtr dict = Runtime.PyImport_GetModuleDict();
-            IntPtr mod = Runtime.IsPython3 ?
-                Runtime.PyImport_ImportModule("builtins")
-                : Runtime.PyDict_GetItemString(dict, "__builtin__");
-            Runtime.PyObject_SetAttrString(mod, "__import__", py_import);
 
+            RestoreImport();
+
+            bool shouldFreeDef = Runtime.Refcount(py_clr_module) == 1;
             Runtime.XDecref(py_clr_module);
             py_clr_module = IntPtr.Zero;
+#if PYTHON3
+            if (shouldFreeDef)
+            {
+                ReleaseModuleDef();
+            }
+#endif
 
             Runtime.XDecref(root.pyHandle);
             root = null;
-
-            hook.Release();
-            hook = null;
-
-            Runtime.XDecref(py_import);
-            py_import = IntPtr.Zero;
-
             CLRModule.Reset();
         }